Implement authentication features for login and logout
Added the 'next-auth' package in the package.json, and used it to create a new authentication configuration file, established the login and logout functionalities, and built a new login page. A middleware file was also created to assist in accessing resources. Modified the 'LoginForm' to dispatch error messages if the user submits invalid email or password. An 'actions' file is created, which includes a new authenticate function to handle the successful and failed login attempts, and handle different types of login errors.